// src/cnab/enums/favorecido-empresa.enum.ts
export enum FavorecidoEmpresaNomeEnum {
  VLT = 'Concessionária do VLT Carioca S.A.',
  Intersul = 'Consórcio Intersul Transportes',
  Internorte = 'Consórcio Internorte de Transportes',
  Transcarioca = 'Consórcio Transcarioca de Transportes',
  SantaCruz = 'Consórcio Santa Cruz Transportes',
  CMTC = 'Companhia Municipal de Transportes Coletivos CMTC Rio',
}

export enum FavorecidoEmpresaCpfCnpjEnum {
  VLT = '18201378000119',
  Intersul = '12464869000176',
  Internorte = '12464539000180',
  Transcarioca = '12464553000184',
  SantaCruz = '12464577000133',
  CMTC = '44520687000161',
}